requirements.txt
pandas
openpyxl
jieba
scikit-learn
gradio
pyyaml
